requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pillow>=10.0.0
qrcode>=7.4.0
python-barcode>=0.15.0
pyserial>=3.5
pyusb>=1.2.0
